split `src/schemas/` into `src/diagnostics/` and `src/documentParsing/` for accuracy; update imports; update test structures and naming
