+++ src/diagnostics/constants.ts
