+++ src/quickpicks/utils/schemas.ts
@@ -1,10 +1,10 @@
+import { SubjectNameStrategy } from "../../diagnostics/produceMessage";
 import { ALLOW_OLDER_SCHEMA_VERSIONS } from "../../extensionSettings/constants";
 import { ResourceLoader } from "../../loaders";
 import { Schema } from "../../models/schema";
 import { SchemaRegistry } from "../../models/schemaRegistry";
 import { KafkaTopic } from "../../models/topic";
 import { showErrorNotificationWithButtons } from "../../notifications";
-import { SubjectNameStrategy } from "../../schemas/produceMessageSchema";
 import { schemaVersionQuickPick } from "../schemas";
 import { getSubjectNameForStrategy } from "./schemaSubjects";
